Renamed `runtime::runtime` to `runtime::scoped_runtime` because `clippy` complained about `clippy::module_inception`.
